requirements.txt
streamlit>=1.20.0
python-dotenv>=0.21.0
mapbox>=0.18.1
geopy>=2.2.0
pandas>=1.5.0
openpyxl>=3.0.0
